[iOS] chore: PR 리뷰 반영 (#53)
conflict 해결 및 디테일 padding 도 조금 수정했습니다.
