+++ Mondee/Model/TabModel.swift
@@ -1,0 +1,21 @@
+//
+//  TabModel.swift
+//  Mondee
+//
+//  Created by hyunjun on 2023/07/20.
+//
+
+import SwiftUI
+
+struct TabItem: Identifiable {
+    var id = UUID()
+    var text: String
+    var icon: String
+    var tab: Tab
+}
+
+enum Tab: String {
+    case today
+    case record
+    case cleanRoom
+}
